Add validation for users&cards

+++ controllers/users.js
@@ -1,18 +1,27 @@
 /* eslint-disable import/no-unresolved */
+const { ObjectId } = require('mongodb');
 const User = require('../models/user');
 
 //отобразить всех пользователей
 const getUsers = (req, res) => {
   User.find({})
-      .then((users) => res.send({ data: users }))
-      .catch((err) => res.status(500).send({ message: err.message }));
+      .then(users => res.send({ data: users }))
+      .catch(err => res.status(500).send({ message: err.message }));
 };
 
 //найти пользователя по id
 const getUserById = (req, res) => {
-  User.findById(req.params.id)
-      .then((user) => { res.send({ data: user }); })
-      .catch((err) => res.status(404).send({ message: `Нет пользователя с таким id`, error: err.message })); //в задании указано 404 статус
+  if (ObjectId.isValid(req.params.id)) {
+    User.findById(req.params.id)
+    .then(user => {
+      if (!user) {
+        return res.status(404).send({ message: `Пользователя с id: ${req.params.id} не существует`})
+      }
+      res.send({ data: user }); })
+    .catch(err => res.status(500).send({ message: `Произошла ошибка при запросе пользователя с id: ${req.params.id}`, error: err.message }));
+  } else {
+    res.status(400).send({ message: 'Неправильный формат id пользователя' });
+  }
 };
 
 //создать пользователя
@@ -20,33 +29,57 @@ const createUser = (req, res) => {
   const { name, about, avatar } = req.body;
 
   User.create({ name, about, avatar })
-      .then((user) => res.status(201).send({ data: user }))
-      .catch((err) => res.status(500).send({ message: `Произошла ошибка при создании пользователя`, error: err.message }));
+      .then(user => res.status(201).send({ data: user }))
+      .catch(err => res.status(500).send({ message: `Произошла ошибка при создании пользователя`, error: err.message }));
 };
 
 //обновить данные пользователя
 const updateUser = (req, res) => {
   const { name, about } = req.body;
 
-  User.findByIdAndUpdate(req.user._id, { name, about }, { new: true, runValidators: true })
-      .then(user => res.send({ data: user }))
-      .catch(err => res.status(400).send({ message: `Нет пользователя с таким id`, error: err.message }));
+  if (ObjectId.isValid(req.params.id)) {
+    User.findByIdAndUpdate(req.params.id, { name, about }, { new: true, runValidators: true })
+    .then(user => {
+      if (!user) {
+        return res.status(404).send({ message: `Пользователя с id: ${req.params.id} не существует`})
+      }
+      res.send({ data: user }); })
+    .catch(err => res.status(500).send({ message: `Произошла ошибка при обновлении пользователя с id: ${req.params.id}`, error: err.message }));
+  } else {
+    res.status(400).send({ message: 'Неправильный формат id пользователя' });
+  }
 };
 
 //обновить аватарку пользователя
 const updateUserAvatar = (req, res) => {
   const { avatar } = req.body;
 
-  User.findByIdAndUpdate(req.user._id, { avatar }, { new: true, runValidators: true })
-      .then(user => res.send({ data: user }))
-      .catch(err => res.status(400).send({ message: `Нет пользователя с таким id`, error: err.message }));
+  if (ObjectId.isValid(req.params.id)) {
+    User.findByIdAndUpdate(req.params.id, { avatar }, { new: true, runValidators: true })
+    .then(user => {
+      if (!user) {
+        return res.status(404).send({ message: `Пользователя с id: ${req.params.id} не существует`})
+      }
+      res.send({ data: user }); })
+    .catch(err => res.status(500).send({ message: `Произошла ошибка при обновлении аватара пользователя с id: ${req.params.id}`, error: err.message }));
+  } else {
+    res.status(400).send({ message: 'Неправильный формат id пользователя' });
+  }
 };
 
 //удалить пользователя
 const deleteUser = (req, res) => {
-  User.findByIdAndDelete(req.params.id)
-      .then((user) => res.send({ message: `Пользователь успешно удалён: ${user.name}` }))
-      .catch((err) => res.status(500).send({ message: `Нет пользователя с таким id`, error: err.message }));
+  if (ObjectId.isValid(req.params.id)) {
+    User.findByIdAndDelete(req.params.id)
+    .then(user => {
+      if (!user) {
+        return res.status(404).send({ message: `Пользователя с id: ${req.params.id} не существует`})
+      }
+      res.send({ message: `Пользователь успешно удалён: ${user.name}` }); })
+    .catch(err => res.status(500).send({ message: `Произошла ошибка при удалении пользователя с id: ${req.params.id}`, error: err.message }));
+  } else {
+    res.status(400).send({ message: 'Неправильный формат id пользователя' });
+  }
 };
 
 module.exports = {
